refactor: 페이지명 변경
