+++ src/constants/useDetailPage.ts
